requirements.txt
django==5.0
pillow==10.1.0
reportlab==4.0.7
python-docx==1.0.1